+++ decidim-bulletin_board-ruby/lib/decidim/bulletin_board/client.rb
@@ -1,40 +1,33 @@
 # frozen_string_literal: true
 
 require "decidim/bulletin_board/command"
+require "decidim/bulletin_board/graphql/factory"
+require "decidim/bulletin_board/settings"
+
+require "decidim/bulletin_board/authority/create_election"
+require "decidim/bulletin_board/authority/end_vote"
+require "decidim/bulletin_board/authority/get_election_status"
+require "decidim/bulletin_board/authority/start_key_ceremony"
+require "decidim/bulletin_board/authority/start_tally"
+require "decidim/bulletin_board/authority/start_vote"
+require "decidim/bulletin_board/authority/publish_results"
+require "decidim/bulletin_board/authority/get_election_results"
+require "decidim/bulletin_board/voter/cast_vote"
+require "decidim/bulletin_board/voter/get_pending_message_status"
 
 module Decidim
   module BulletinBoard
     # The Bulletin Board client
     class Client
-      def initialize
-        @server = BulletinBoard.server.presence
-        @api_key = BulletinBoard.api_key.presence
-        @scheme = BulletinBoard.scheme.presence
-        @authority_name = BulletinBoard.authority_name.presence
-        @number_of_trustees = BulletinBoard.number_of_trustees.presence
-        @server_public_key = BulletinBoard.server_public_key.presence
-        @identification_private_key = BulletinBoard.identification_private_key.presence
-        @private_key = identification_private_key_content if identification_private_key
+      def initialize(config = Decidim::BulletinBoard)
+        @settings = Settings.new(config)
+        @graphql = Graphql::Factory.client_for(settings)
       end
 
-      attr_reader :server, :scheme, :api_key, :number_of_trustees, :authority_name
-
-      delegate :authority_slug, :server_public_key, to: Decidim::BulletinBoard::Command
-
-      def quorum
-        @scheme.dig(:parameters, :quorum) || number_of_trustees
-      end
-
-      def public_key
-        private_key&.export
-      end
-
-      def configured?
-        private_key && server && api_key
-      end
+      delegate :configured?, :server, :public_key, :authority_name, :number_of_trustees, :quorum, to: :settings
 
       def create_election(election_id, election_data)
-        create_election = Decidim::BulletinBoard::Authority::CreateElection.new(election_id, election_data)
+        create_election = configure Authority::CreateElection.new(election_id, election_data)
         yield create_election.message_id if block_given?
         create_election.on(:ok) { |election| return election }
         create_election.on(:error) { |error_message| raise StandardError, error_message }
@@ -42,7 +35,7 @@ module Decidim
       end
 
       def start_key_ceremony(election_id)
-        start_key_ceremony = Decidim::BulletinBoard::Authority::StartKeyCeremony.new(election_id)
+        start_key_ceremony = configure Authority::StartKeyCeremony.new(election_id)
         yield start_key_ceremony.message_id if block_given?
         start_key_ceremony.on(:ok) { |pending_message| return pending_message }
         start_key_ceremony.on(:error) { |error_message| raise StandardError, error_message }
@@ -50,7 +43,7 @@ module Decidim
       end
 
       def start_vote(election_id)
-        start_vote = Decidim::BulletinBoard::Authority::StartVote.new(election_id)
+        start_vote = configure Authority::StartVote.new(election_id)
         yield start_vote.message_id if block_given?
         start_vote.on(:ok) { |pending_message| return pending_message }
         start_vote.on(:error) { |error_message| raise StandardError, error_message }
@@ -58,7 +51,7 @@ module Decidim
       end
 
       def cast_vote(election_id, voter_id, encrypted_vote)
-        cast_vote = Decidim::BulletinBoard::Voter::CastVote.new(election_id, voter_id, encrypted_vote)
+        cast_vote = configure Voter::CastVote.new(election_id, voter_id, encrypted_vote)
         yield cast_vote.message_id if block_given?
         cast_vote.on(:ok) { |pending_message| return pending_message }
         cast_vote.on(:error) { |error_message| raise StandardError, error_message }
@@ -66,14 +59,14 @@ module Decidim
       end
 
       def get_pending_message_status(message_id)
-        get_pending_message_status = Decidim::BulletinBoard::Voter::GetPendingMessageStatus.new(message_id)
+        get_pending_message_status = configure Voter::GetPendingMessageStatus.new(message_id)
         get_pending_message_status.on(:ok) { |status| return status }
         get_pending_message_status.on(:error) { |error_message| raise StandardError, error_message }
         get_pending_message_status.call
       end
 
       def end_vote(election_id)
-        end_vote = Decidim::BulletinBoard::Authority::EndVote.new(election_id)
+        end_vote = configure Authority::EndVote.new(election_id)
         yield end_vote.message_id if block_given?
         end_vote.on(:ok) { |pending_message| return pending_message }
         end_vote.on(:error) { |error_message| raise StandardError, error_message }
@@ -81,14 +74,14 @@ module Decidim
       end
 
       def get_election_status(election_id)
-        get_election_status = Decidim::BulletinBoard::Authority::GetElectionStatus.new(election_id)
+        get_election_status = configure Authority::GetElectionStatus.new(election_id)
         get_election_status.on(:ok) { |status| return status }
         get_election_status.on(:error) { |error_message| raise StandardError, error_message }
         get_election_status.call
       end
 
       def start_tally(election_id)
-        start_tally = Decidim::BulletinBoard::Authority::StartTally.new(election_id)
+        start_tally = configure Authority::StartTally.new(election_id)
         yield start_tally.message_id if block_given?
         start_tally.on(:ok) { |pending_message| return pending_message }
         start_tally.on(:error) { |error_message| raise StandardError, error_message }
@@ -96,14 +89,14 @@ module Decidim
       end
 
       def get_election_results(election_id)
-        get_log_entries = Decidim::BulletinBoard::Authority::GetElectionResults.new(election_id)
+        get_log_entries = configure Authority::GetElectionResults.new(election_id)
         get_log_entries.on(:ok) { |result| return result }
         get_log_entries.on(:error) { |error_message| raise StandardError, error_message }
         get_log_entries.call
       end
 
       def publish_results(election_id)
-        publish_results = Decidim::BulletinBoard::Authority::PublishResults.new(election_id)
+        publish_results = configure Authority::PublishResults.new(election_id)
         yield publish_results.message_id if block_given?
         publish_results.on(:ok) { |status| return status }
         publish_results.on(:error) { |error_message| raise StandardError, error_message }
@@ -112,10 +105,11 @@ module Decidim
 
       private
 
-      attr_reader :identification_private_key, :private_key
+      attr_reader :settings, :graphql
 
-      def identification_private_key_content
-        @identification_private_key_content ||= JwkUtils.import_private_key(identification_private_key)
+      def configure(command)
+        command.configure(settings, graphql)
+        command
       end
     end
   end