+++ decidim-bulletin_board-ruby/spec/decidim/bulletin_board/command_spec.rb
@@ -6,12 +6,12 @@ require "wisper/rspec/stub_wisper_publisher"
 module Decidim
   module BulletinBoard
     describe Command do
-      subject(:instance) { described_class.new }
+      subject(:command) { described_class.new }
 
       include_context "with a configured bulletin board"
 
       describe ".build_message_id" do
-        subject { instance.build_message_id(unique_election_id, type_subtype, voter_id) }
+        subject { command.build_message_id(unique_election_id, type_subtype, voter_id) }
 
         let(:unique_election_id) { "decidim-test-authority.26" }
         let(:type_subtype) { "vote.cast" }
@@ -31,7 +31,7 @@ module Decidim
       end
 
       describe ".complete_message" do
-        subject { instance.complete_message(message_id, message) }
+        subject { command.complete_message(message_id, message) }
 
         let(:message_id) { "a.message.id" }
         let(:message) { { a: "message", with: "data" } }
@@ -50,7 +50,7 @@ module Decidim
       end
 
       describe ".sign_message" do
-        subject { instance.sign_message(message_id, message) }
+        subject { command.sign_message(message_id, message) }
 
         let(:message_id) { "a.message.id" }
         let(:message) { { a: "message", with: "data" } }
@@ -60,25 +60,6 @@ module Decidim
           subject
           expect(JWT).to have_received(:encode).with({ a: "message", iat: Time.now.to_i, message_id: "a.message.id", with: "data" }, instance_of(OpenSSL::PKey::RSA), "RS256")
         end
-      end
-
-      TEST_PUBLIC_KEY = "-----BEGIN PUBLIC KEY-----\nMIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEApNgMt8lnPDD3TlWYGhRi\nV1oZkPQmnLdiUzwyb/+35qKD9k+HU86xo0uSgoOUWkBtnvFscq8zNDPAGAlZVoka\nN/z9ksZblSce0LEl8lJa3ICgghg7e8vg/7Lz5dyHSQ3PCLgenyFGcL401aglDde1\nXo4ujdz33Lklc4U9zoyoLUI2/viYmNOU6n5Mn0sJd30FeICMrLD2gX46pGe3MGug\n6groT9EvpKcdOoJHKoO5yGSVaeY5+Bo3gngvlgjlS2mfwjCtF4NYwIQSd2al+p4B\nKnuYAVKRSgr8rYnnjhWfJ4GsCaqiyXNi5NPYRV6gl/cx/1jUcA1rRJqQR32I8c8Q\nbAXm5qNO4URcdaKys9tNcVgXBL1FsSdbrLVVFWen1tfWNfHm+8BjiWCWD79+uk5g\nI0SjC9tWvTzVvswWXI5weNqqVXqpDydr46AsHE2sG40HRCR3UF3LupT+HwXTcYcO\nZr5dJClJIsU3Hrvy4wLssub69YSNR1Jxn+KX2vUc06xY8CNIuSMpfufEq5cZopL6\nO2l1pRsW1FQnF3s078/Y9MaQ1gPyBo0IipLBVUj5IjEIfPuiEk4jxkiUYDeqzf7b\nAvSFckp94yLkRWTs/pEZs7b/ogwRG6WMHjtcaNYe4CufhIm9ekkKDeAWOPRTHfKN\nmohRBh09XuvSjqrx5Z7rqb8CAwEAAQ==\n-----END PUBLIC KEY-----\n"
-      describe "#private_key" do
-        subject { described_class.private_key }
-
-        it { expect(subject.public_key.to_s).to eq(TEST_PUBLIC_KEY) }
-      end
-
-      describe "#authority_slug" do
-        subject { described_class.authority_slug }
-
-        it { expect(subject).to eq("decidim-test-authority") }
-      end
-
-      describe "#server_public_key_rsa" do
-        subject { described_class.server_public_key_rsa }
-
-        it { expect(subject).to be_a OpenSSL::PKey::RSA }
       end
     end
   end